[refactor] удалил не нужные папки
