+++ frontend/src/styles/_typography.scss
@@ -10,7 +10,7 @@ $headline1-line-height: 4rem;
 $headline1-font-weight: $font-weight-regular;
 $headline1-letter-spacing: 0;
 
-// Headline 2 (Large / pxphasized)
+// Headline 2 (Large / emphasized)
 $headline2-font-size: 3.2rem;
 $headline2-line-height: 4rem;
 $headline2-font-weight: $font-weight-medium;
@@ -28,13 +28,13 @@ $title1-small-line-height: 2rem;
 $title1-small-font-weight: $font-weight-medium;
 $title1-small-letter-spacing: 0.1px;
 
-// Title 2 (Medium / pxphasized)
+// Title 2 (Medium / emphasized)
 $title2-medium-font-size: 1.8rem;
 $title2-medium-line-height: 2.4rem;
 $title2-medium-font-weight: $font-weight-bold;
 $title2-medium-letter-spacing: 0.15px;
 
-// Title 2 (Small / pxphasized)
+// Title 2 (Small / emphasized)
 $title2-small-font-size: 1.4rem;
 $title2-small-line-height: 2rem;
 $title2-small-font-weight: $font-weight-bold;
@@ -46,7 +46,7 @@ $label1-large-line-height: 2rem;
 $label1-large-font-weight: $font-weight-medium;
 $label1-large-letter-spacing: 0.1px;
 
-// Label 2 (Large / pxphasized)
+// Label 2 (Large / emphasized)
 $label2-large-font-size: 1.4rem;
 $label2-large-line-height: 2rem;
 $label2-large-font-weight: $font-weight-bold;
@@ -58,7 +58,7 @@ $body1-medium-line-height: 2rem;
 $body1-medium-font-weight: $font-weight-regular;
 $body1-medium-letter-spacing: 0.25px;
 
-// Body 2 (Medium / pxphasized)
+// Body 2 (Medium / emphasized)
 $body2-medium-font-size: 1.4rem;
 $body2-medium-line-height: 2rem;
 $body2-medium-font-weight: $font-weight-medium;
